Split definition of tridiagonal matrix
